requirements.txt
fastapi
uvicorn[standard]
diffusers>=0.30.0
transformers
accelerate
pillow
python-multipart
sentencepiece
opencv-python
imageio 
imageio-ffmpeg
insightface
controlnet-aux